Import specified publishers and create sample submissions
Contracts Finder notices in CSV form are collected in a source directory.
They are aggregated, filtered for publisher names, and collected in new
CSV files specific to the weekly time period and notice type, and to the publisher
if the `publisher_submissions` flag is set. These CSV files are turned
into individual release package files. They are loaded into the database
if the `load_data` flag is set.